Refactored Geometry/Objects (long -> double) and added some functions, added printing doubles to FastOutput (should add setting precision later)

+++ Geometry/ClosestPairEuclidean.java
@@ -4,11 +4,11 @@ class ClosestPairEuclidean {
         final int n = pts.length;
         Arrays.sort(pts, (a, b) -> comp(a, b));
         final TreeSet<Point> s = new TreeSet<>((a, b) -> comp(a, b));
-        long bestDist = Long.MAX_VALUE;
+        double bestDist = Double.MAX_VALUE;
         final Point[] result = new Point[2];
         int j = 0;
         for(int i = 0; i < n; ++i){
-            long d = (long)Math.ceil(Math.sqrt(bestDist));
+            double d = (double)Math.ceil(Math.sqrt(bestDist));
             while(j < n && pts[i].x - pts[j].x >= d){
                 s.remove(new Point(pts[j].y, pts[j].x, 0));
                 ++j;
@@ -16,7 +16,7 @@ class ClosestPairEuclidean {
             Point p1 = s.ceiling(new Point(pts[i].y - d, pts[i].x, 0));
             final Point p2 = s.higher(new Point(pts[i].y + d, pts[i].x, 0));
             while(p1 != null && (p2 == null || p1.x != p2.x || p1.y != p2.y)){
-                long dist = square(pts[i].x - p1.y) + square(pts[i].y - p1.x);
+                double dist = square(pts[i].x - p1.y) + square(pts[i].y - p1.x);
                 if(dist < bestDist){
                     bestDist = dist;
                     result[0] = pts[i];
@@ -28,6 +28,6 @@ class ClosestPairEuclidean {
         }
         return result;
     }
-    private static final int comp(Point a, Point b){ return a.x == b.x ? Long.compare(a.y, b.y) : Long.compare(a.x, b.x); }
-    private static final long square(long x) { return x * x; } 
+    private static final int comp(Point a, Point b){ return a.x == b.x ? Double.compare(a.y, b.y) : Double.compare(a.x, b.x); }
+    private static final double square(double x) { return x * x; } 
 }
